feat: Implement target sequence generation and enable encoder input.
This commit implements the generation of the target sequence using the BUILD_TARGET_SEQUENCE_FOR_LINE_BATCH_SIZE macro.  It also re-enables the encoder input section, although the actual encoder forward pass is still commented out.  Additionally, the data type for the target sequence has been changed to double to match the input sequence.

Key changes:

- Implemented BUILD_TARGET_SEQUENCE_FOR_LINE_BATCH_SIZE and integrated it into the main processing loop.

- Changed the target sequence (	s) data type to double for consistency with the input sequence (is).

- Added print statements to verify the shapes of the input and target sequences.

- Re-enabled the encoder input section (commented out the actual encoder forward call).

- Updated main.cpp to declare 	argetSequence as Collective<double>.

This commit prepares the ground for fully integrating the encoder component and enables end-to-end testing of the input and target sequence generation. The next step will be to uncomment and properly configure the encoder forward pass.

+++ usage/src/main.cpp
@@ -134,7 +134,7 @@ int main(int argc, char* argv[])
     Collective<double> inputSequence;
     Collective<float> position;
     Collective<float> positionEncoding;
-    Collective<float> targetSequence;
+    Collective<double> targetSequence;
     Collective<double> W1 = Collective<double>{NULL, DIMENSIONS{SKIP_GRAM_EMBEDDNG_VECTOR_SIZE, input_sequence_vocab.numberOfTokens(), NULL, NULL}}; 
 
     try
